Add user command tests
