+++ helium_commander/_version.py
@@ -1,5 +1,5 @@
 
 # This file is automatically generated by setup.py.
-__version__ = '0.9.3.post8'
-__sha__ = 'g1940e2c'
-__revision__ = 'g1940e2c'
+__version__ = '0.9.3.post9'
+__sha__ = 'ga82e2e7'
+__revision__ = 'ga82e2e7'
