+++ tests/cassettes/tests.commands.test_user.test_list.json
@@ -1,0 +1,48 @@
+{
+  "http_interactions": [
+    {
+      "recorded_at": "2016-09-08T17:22:38",
+      "request": {
+        "body": {
+          "encoding": "utf-8",
+          "string": ""
+        },
+        "headers": {
+          "Accept": "application/json",
+          "Accept-Charset": "utf-8",
+          "Accept-Encoding": "gzip, deflate",
+          "Authorization": "<AUTH_TOKEN>",
+          "Connection": "keep-alive",
+          "Content-Type": "application/json",
+          "User-Agent": "helium-python/0.0.6"
+        },
+        "method": "GET",
+        "uri": "https://api.helium.com/v1/user"
+      },
+      "response": {
+        "body": {
+          "base64_string": "H4sIAAAAAAAEA32OQQqDMBBFr1JmbSSJWjSrFlz0AK7cTeNAA0ZFJ0IR715T6bbLz3883gYdMoLZAJln9wxMS1wDegIDD+pd8DWtd2vHMPClJj/CnoDrjrejVeB5LLfXF03tCAl4Op3k0fX/QDsTMkWXlioXSgpdNVqZrDSFbg9TmI6+H3AVshSqalRpZGG0TLNc5VnVxiB+TzE4LDTDvn8A6qj969cAAAA=",
+          "encoding": null,
+          "string": ""
+        },
+        "headers": {
+          "Access-Control-Allow-Headers": "Origin, Content-Type, Accept, Authorization",
+          "Access-Control-Allow-Origin": "*",
+          "Airship-Quip": "never breaks eye contact",
+          "Airship-Trace": "b13,b12,b11,b10,b09,b08,b07,b06,b05,b04,b03,c03,c04,d04,e05,e06,f06,f07,g07,g08,h10,i12,l13,m16,n16,o16,o17,o18",
+          "Content-Encoding": "gzip",
+          "Content-Type": "application/json",
+          "Date": "Thu, 08 Sep 2016 17:22:39 GMT",
+          "Server": "Warp/3.2.7",
+          "Transfer-Encoding": "chunked"
+        },
+        "status": {
+          "code": 200,
+          "message": "OK"
+        },
+        "url": "https://api.helium.com/v1/user"
+      }
+    }
+  ],
+  "recorded_with": "betamax/0.8.0"
+}
