+++ tests/commands/test_user.py
@@ -1,0 +1,6 @@
+from .util import cli_run
+
+
+def test_list(client, authorized_user):
+    output = cli_run(['user', 'list'])
+    assert authorized_user.short_id in output
